1、重撸价格显示逻辑 2、修改钟点房priceType显示错误的bug 3、删除测试数据

+++ app/src/main/java/com/huicheng/hotel/android/net/bean/RoomListInfoBean.java
@@ -7,7 +7,7 @@ import java.io.Serializable;
  * @date 2017/4/21 0021
  */
 public class RoomListInfoBean implements Serializable{
-    public float clockPrice;
+    public int clockPrice;
     public boolean hourSupport;
     public int id;
     public String isVip;
